+++ Samples/BasicRendering/Helper/MayaObj.h
@@ -23,15 +23,7 @@ namespace DontUse
       ezVec2 tex0;
     };
 
-    ~MayaObj();
-
     static MayaObj* LoadFromFile(const char* szPath, ezGALDevice* pDevice);
-
-    //ezGALBufferHandle GetVB() const { return m_hVB; }
-
-    //ezGALBufferHandle GetIB() const { return m_hIB; }
-
-    //ezUInt32 GetPrimitiveCount() const { return m_uiPrimitiveCount; }
 
     ezMeshBufferResourceHandle m_hMeshBuffer;
 
@@ -39,15 +31,6 @@ namespace DontUse
 
     MayaObj(const ezArrayPtr<Vertex>& pVertices, const ezArrayPtr<ezUInt16>& pIndices, ezGALDevice* pDevice);
 
-  private:
-
-    //ezGALBufferHandle m_hVB;
-
-    //ezGALBufferHandle m_hIB;
-
-    //ezGALDevice* m_pDevice;
-
-    //ezUInt32 m_uiPrimitiveCount;
 
   };
 }