+++ Samples/BasicRendering/Helper/Misc.h
@@ -1,6 +1,0 @@
-#pragma once
-
-namespace DontUse
-{
-  ezResult ReadCompleteFile(const char* szFile, ezStringBuilder& out_FileContent);
-};